reverting file changes for MultiRowParallelScanInterpreter.java; moving all new classes to src/test
