+++ src/test/java/com/amazon/janusgraph/AbstractGraph.java
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package com.amazon.janusgraph.example;
+package com.amazon.janusgraph;
 
 import com.amazon.janusgraph.triple.Triple;
 import org.janusgraph.core.JanusGraph;
