+++ src/test/java/com/amazon/janusgraph/ScenarioTests.java
@@ -44,7 +44,6 @@ import org.junit.Test;
 
 import com.amazon.janusgraph.diskstorage.dynamodb.BackendDataModel;
 import com.amazon.janusgraph.diskstorage.dynamodb.DynamoDBStoreTransaction;
-import com.amazon.janusgraph.example.TravelGraphFactory;
 import com.google.common.base.Stopwatch;
 
 /**
