+++ src/test/java/com/amazon/janusgraph/TravelGraph.java
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package com.amazon.janusgraph.example;
+package com.amazon.janusgraph;
 
 import java.io.BufferedReader;
 import java.io.IOException;
