+++ src/test/java/com/amazon/janusgraph/triple/Triple.java
